Support non-overlapping application distribution nodes
Currently all known nodes should have the same value for the
kernel application's 'distributed' environment variable.  It
is not expected that any application will be distributed on
on more than one set of nodes.

It should be possible to distributed an application between
multiple non-overlapping sets of nodes.  For example with this
system configuration file on nodes a@a and b@b:

   [{kernel,
      [{distributed, [{app_no, [a@a, b@b]}]},
       {sync_nodes_optional, [a@a, b@b]},
       {sync_nodes_timeout, 5000}]}].

... and this system configuration file on nodes c@c and d@d:

   [{kernel,
      [{distributed, [{app_no, [c@c, d@d]}]},
       {sync_nodes_optional, [c@c, d@d]},
       {sync_nodes_timeout, 5000}]}].

Other applications may be distributed involving some other
combination of these nodes without interference.

This patch adds checks in dist_ac to ignore DAC protocol
messages of an application from nodes not included in that
application's distribution specification locally.